deleted ControllerTest file and moved contents into ControllerTests.java file
