+++ src/tests/ControllerTests.java
@@ -1,5 +1,6 @@
 package tests;
 
+import static org.junit.Assert.assertFalse;
 import static org.junit.jupiter.api.Assertions.*;
 
 import org.junit.jupiter.api.Test;
@@ -19,6 +20,15 @@ class ControllerTests {
 		String result = testController.searchByKeyword(testKeyword);
 		boolean isNull = result.isEmpty();
 		assertFalse(isNull);
+	}
+	
+	@Test
+	public void searchByIngredientsTest() {
+		Controller testController = new Controller();
+		String testIngredient = "banana"; //random ingredient to search with
+		String result = testController.searchByIngredient(testIngredient);
+		boolean IsNull = result.isEmpty();
+		assertFalse(IsNull);
 	}
 
 }
